scientificdefaults_internal.cpp
